+++ diventi/ebooks/views.py
@@ -8,7 +8,11 @@ from django.utils.translation import gettext_lazy as _
 from django.contrib.auth.mixins import LoginRequiredMixin, UserPassesTestMixin
 from django.http import Http404
 
+from dal import autocomplete
+
+from diventi.core.views import StaffRequiredMixin
 from diventi.products.models import Product
+
 from .models import Book, Chapter, Section
 
 
@@ -163,3 +167,16 @@ class SectionDetailView(LoginRequiredMixin, UserHasProductMixin,
         book_slug = self.kwargs.get('book_slug', None)
         book = get_object_or_404(Book, slug=book_slug)
         return ['ebooks/section_detail_%s.html' % book.template]
+
+
+class ChapterAutocomplete(autocomplete.Select2QuerySetView, StaffRequiredMixin):
+    """ Returns a list of chapters to facilitate user form fill. """
+  
+    def get_queryset(self):
+        qs = Chapter.objects.all()
+        book = self.forwarded.get('book', None)
+        if book:
+            qs = qs.filter(chapter_book=book)
+        if self.q:
+            qs = qs.filter(title__istartswith=self.q)
+        return qs