+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - [core] Upgraded material kit from v2.1.1 to v2.2.0
 - [core] Upgraded django from v2.2.9 to v2.2.10
 - [core] Added support for Opengraph on blog, landing and product pages
+- [ebooks] Added support for chapter autocompletion based on book's titles in the admin page
 - [feedbacks] Introduced an email notification for the authors when a user completes a survey
 - [feedbacks] Introduced created and modified dates to answers
 - [landing] Introduced 'cards section': a new section template
@@ -20,6 +21,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 - [accounts] Updated user collection in order to separate authored products from those bought
+- [core] Enabled callout visual styles for the ckeditor
 - [landing] Changed the ordering of search results to make products appear first
 - [landing] Updated dashboard visual styles with rounded buttons and card-styled sections 
 - [products] Enabled product details and colors on detail page